refactor(es): Extract legacy flag accessors into config_legacy.go
Move all code dealing with deprecated rotation flags into a dedicated
file: feature gate, getters, hasAnyLegacyRotationFlags, resolvedRotation,
and LogDeprecationWarnings. Make alias getters unexported since they are
only used within the config package (by resolvedRotation and Validate).
